fix a few bugs and add train_all.py

+++ hybrid_graph/cli.py
@@ -36,7 +36,7 @@ class Main:
             'type': float, 'default': 0.01, 'help': 'Initial learning rate.',
         },
         ('-m', '--max-epochs'): {
-            'type': float, 'default': 100,
+            'type': int, 'default': 100,
             'help': 'Maximum number of epochs for training.',
         },
         ('-b', '--batch-size'): {
